lib/docker/requirements.txt
PyYAML
click
click-default-group
python-slugify
boto3
playwright
shot-scraper